requirement.txt
django
djangorestframework
django-cors-headers
celery
drf-yasg
django-environ
mysqlclient